# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
            "api.ChangePasswordRequest": {
                "properties": {
                    "current_password": {
                        "type": "string"
                    },
                    "new_password": {
                        "type": "string"
                    }
                },
                "required": [
                    "current_password",
                    "new_password"
                ],
                "type": "object"
            },
...
            "api.CurrentUser": {
                "properties": {
                    "role": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    }
                },
                "required": [
                    "role",
                    "username"
                ],
                "type": "object"
            },
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "api.LoginResponse": {
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "required": [
                    "token"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/me": {
            "get": {
                "description": "Read the authenticated account name and role. Viewer accounts can inspect resources; administrator access is required for VM and host mutations.",
                "operationId": "currentUser",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.CurrentUser"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "currentUser",
                "tags": [
                    "login"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            }
        },
...
        "/api/me/password": {
            "patch": {
                "description": "Change the current account password by supplying current_password and new_password. Existing login sessions may become invalid. The response contains a new login token; treat it as a secret.",
                "operationId": "changePassword",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.ChangePasswordRequest",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.LoginResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "changePassword",
                "tags": [
                    "login"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            }
        },
...
        "/api/vms/{id}/console": {
            "get": {
                "description": "Interactive VM serial console over WebSocket. Not exposed as a request/response MCP tool.",
                "operationId": "consoleVM",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "101": {
                        "description": "WebSocket upgrade"
                    }
                },
                "summary": "consoleVM",
                "tags": [
                    "vms"
                ],
                "x-maco": {
                    "expose": false,
                    "readOnly": true,
                    "transport": "unsupported"
                }
            }
        },
...
        "/api/vms/{id}/display": {
            "get": {
                "description": "Interactive VM graphical display over WebSocket. Not exposed as a request/response MCP tool.",
                "operationId": "displayVM",
                "parameters": [
                    {
                        "description": "id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "101": {
                        "description": "WebSocket upgrade"
                    }
                },
                "summary": "displayVM",
                "tags": [
                    "vms"
                ],
                "x-maco": {
                    "expose": false,
                    "readOnly": true,
                    "transport": "unsupported"
                }
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ],
...
}